Malloc works.
Problems:
Userspace is concurrent, and malloc uses static data. Ergo, it's not process safe (virtual mem isn't on yet).
Still need to free mem on cleanup (and make a global memory system that can do something with free memory ;))

+++ malloc.h
@@ -5,8 +5,10 @@
 #error __FILE__ " included from kernel code."
 #endif
 
+#include "types.h"
+
 int mem_init(size_t pages);
-void* malloc(int size);
+void* malloc(size_t size);
 int free(void *ptr);
 void Mem_Dump();
